==> 0_get_youtubers.ipynb <==
#scraping da
#https://blog.feedspot.com/american_makeup_youtubers/
#per prendere i canali aggiornati
#prendo i top 60

import pandas as pd

import requests 

from bs4 import BeautifulSoup

import os

import json

import googleapiclient.discovery
import googleapiclient.errors

    request = youtube.search().list(
        part="snippet",
        maxResults=50,
        order="viewCount",
        q="makeup",
        regionCode="US",
        type="channel"
    )
    response = request.execute()

    request = youtube.search().list(
        part="snippet",
        maxResults=50,
        order="viewCount",
        q="makeup -topic",
        regionCode="US",
        relevanceLanguage="en",
        type="channel"
    )

chiave_yt = json.load(open('key.json', 'r'))['youtube']['key']

youtube = googleapiclient.discovery.build('youtube', 'v3', developerKey=chiave_yt)

url = 'https://blog.feedspot.com/american_makeup_youtubers/'

html_soup = BeautifulSoup(html_reply, 'html.parser')

#prendo userid degli influencers (o channelid se userid è assente)
youtubers_id = []
selettore = 'p>a, .ext'
for link in html_soup.select(selettore):
    if(link.get('href').split('/')[0] == 'https:'):
        youtubers_id.append([link.get('href').split('/')[3],link.get('href').split('/')[4]])

#ricavo i channelId mancanti

responses = []
for user in youtubers:
    if user[2] == None:
        request = youtube.channels().list(
            part="snippet",
            forUsername=user[1]
        )
        response = request.execute()
        responses.append(response)
    else:
        request = youtube.channels().list(
        part="snippet",
        id=user[2]
        )
        response = request.execute()
        responses.append(response)

youtubers = []
for r , yt_id in zip(responses,youtubers_id):
    temp_user = []
    #display_name
    temp_user.append(r['items'][0]['snippet']['title'])
    #channel_id
    temp_user.append(r['items'][0]['id'])
    #se c'è userid, allora anche userid
    if yt_id[0]=='user':
        temp_user.append(yt_id[1])
    else:
        temp_user.append(None)
    youtubers.append(temp_user)

df_youtubers = pd.DataFrame(youtubers,columns=['display_name','channel_id','user_id'])

df_youtubers

df_youtubers.to_csv('data/get_youtubers_output/youtubers.csv', index = False)

#per leggere
#df = pd.read_csv('data/get_youtubers_output/youtubers.csv')
