==> 0_train_sentiment_analysis_model.ipynb <==
import tensorflow as tf
from tensorflow import keras
import tensorflow_hub as hub
import tensorflow_datasets as tfds

model = keras.Sequential([
    hub.KerasLayer("https://tfhub.dev/google/tf2-preview/nnlm-en-dim50/1",
                  dtype=tf.string, input_shape=[], output_shape=[50]),
    keras.layers.Dense(128, activation="relu"),
    keras.layers.Dense(1, activation="sigmoid")
])

model.compile(loss = "binary_crossentropy", optimizer="adam", metrics=["accuracy"])

datasets, info = tfds.load("imdb_reviews", as_supervised=True,with_info=True)
train_size=info.splits['train'].num_examples
batch_size = 32
train_set = datasets['train'].batch(batch_size).prefetch(1)

history = model.fit(train_set, epochs=10)

model.save('sentiment_analysis_model.h5')

#model = tf.keras.models.load_model('sentiment_analysis_model.h5',custom_objects={'KerasLayer':hub.KerasLayer})
