==> 1_get_videos.ipynb <==
import pandas as pd

import requests 

import os

import json

import googleapiclient.discovery
import googleapiclient.errors

df_youtubers = pd.read_csv('data/get_youtubers_output/youtubers.csv')

chiave_yt = json.load(open('key.json', 'r'))['youtube']['key']
youtube = googleapiclient.discovery.build('youtube', 'v3', developerKey=chiave_yt)

quotas = 0

def get_videos_id(items):
    ids = []
    for item in items:
        #print(item['contentDetails'])
        try:
            ids.append(item['contentDetails']['upload']['videoId'])
        except KeyError:
            #print('Not upload')
            continue
    return ids

youtuber_videos = []
for idx, youtuber in enumerate(df_youtubers.values):
    print("___________________________________")
    print("{} Youtuber: {}".format(idx, youtuber[0]))
    channel_id = youtuber[1]
    videos = []
    next_page_token = " "
    while True:
        if quotas < 9000:
            request = youtube.activities().list(
                part="contentDetails",
                channelId=channel_id,
                maxResults=50,
                pageToken = next_page_token
            )
            response = request.execute()
            quotas += 3
            ids = get_videos_id(response['items'])
            print('Got {} new videos'.format(len(ids)))
            for vid in ids:
                if vid not in videos:
                    videos.append(vid)
            try:
                next_page_token = response['nextPageToken']
            except KeyError:
                break
            if len(videos) > 24:
                break
        else:
            print("Quota giornaliera superata")
    #print("Channel: {}, videos: {}".format(channel_id, videos))
    for vid in videos:
        youtuber_videos.append([channel_id,vid])

df_youtuber_videos = pd.DataFrame(youtuber_videos,columns=['channel_id','video_id'])

df_youtuber_videos

df_youtuber_videos.to_csv('data/get_youtuber_videos_output/youtuber_videos.csv', index = False)

quotas
