==> 2_1_compute_video_details.ipynb <==
import pandas as pd

import os

import numpy as np

import pickle

import ast

import spacy
from spacy.lang.en.stop_words import STOP_WORDS
import string
nlp = spacy.load("en_core_web_md")

from sklearn.feature_extraction.text import TfidfVectorizer

from IPython.display import clear_output

n_video_details_files = int(np.load('data/get_videos_details/last_csv.npy'))
n_video_details_files

for i in range(n_video_details_files+1):
    if i == 0:
        df_videos_details = pd.read_csv('data/get_videos_details/videos_details_{}.csv'.format(i))
    else:
        df = pd.read_csv('data/get_videos_details/videos_details_{}.csv'.format(i))
        df_videos_details = df_videos_details.append(df,ignore_index=True)
        

df_youtubers = pd.read_csv('data/get_youtubers_output/youtubers.csv')

youtubers = list(df_youtubers['display_name'].values)

topics = dict()
with open('data/youtube_topicIDs.txt', 'r') as f:
    for line in f.readlines():
        key, value = line.split(',')[0], line.split(',')[1][:-1]
        topics.update({key:value})

np.save('data/topic_ids.npy', topics, allow_pickle=True)
#t = np.load('data/topic_ids.npy',allow_pickle=True).tolist()

keywords = []
with open('data/key_words.txt', 'r') as f:
    for line in f.readlines():
        keywords.append(line.strip().lower().split(','))

np.save('data/keywords.npy', keywords, allow_pickle=True)

makeup_brands = []
with open('data/makeup_brands.txt', 'r') as f:
    for line in f.readlines():
        makeup_brands.append(line.strip().lower().split(','))

np.save('data/makeup_brands.npy', makeup_brands, allow_pickle=True)

categories = dict()
with open('data/youtube_categoryIDs.txt', 'r') as f:
    for line in f.readlines():
        key, value = line.split('-')[0].strip(), line.split('-')[1][:-1].strip()
        categories.update({key:value})


np.save('data/categories_ids.npy', categories, allow_pickle=True)
#t = np.load('data/topic_ids.npy',allow_pickle=True).tolist()

def tokenizer(sentence):
    arr = []
    for w in sentence.split(' '):
        arr.append(w.translate(str.maketrans('', '', string.punctuation)).strip().lower())
    return arr

def find_topics(title, relevant_topics):
    title_topics = set()
    other_topics = set()
    for keyword, generic_keyword in keywords:
        if len(keyword.split(' '))>1:
            matches = keyword.strip().split(' ')
            if all(x in title.lower().split(' ') for x in matches):
                title_topics.add(generic_keyword)
        else:
            if keyword in tokenizer(title):
                title_topics.add(generic_keyword)
    if relevant_topics != None:
        for t in relevant_topics:
            if t == 'Lifestyle_(sociology)':
                other_topics.add('lifestyle')
            else:
                other_topics.add(t.lower())
        total_topics = title_topics.union(other_topics)
        return total_topics
    else:
        return title_topics

def find_brands_and_youtubers(description):
    cited_brand = set()
    cited_youtubers = set()
    for brand in makeup_brands:
        original = brand[-1]
        matches = brand[0].split(' ')
        if all(x in description.lower().split(' ') for x in matches):
            cited_brand.add(original)
    for youtuber in youtubers:
        matches = youtuber.lower().split(' ')
        if all(x in description.lower().split(' ') for x in matches):
            cited_youtubers.add(original)
    return cited_brand, cited_youtubers

def compute_details(df):
    df_arr_computed = []
    for idx, row in enumerate(df.values):
        print('Video: {}/{}'.format(idx,len(df.values)))
        clear_output(wait=True)
        relevant_topics_cat=None
        line=[row[0],row[1],row[2]]
        try:
            line.append(categories[str(row[3])])
        except:
            line.append(None)
        try:
            l = row[4]
            l = ast.literal_eval(l)
            l = [n.strip() for n in l]
            tps = [topics[item] for item in l]
            line.append(tps)
        except:
            line.append(None)
        try:
            v = row[5]
            v = ast.literal_eval(v)
            v = [n.split('/')[-1] for n in v]
            relevant_topics_cat = v
            line.append(v)
        except:
            line.append(None)
        #vedo se il titolo contiene keywords
        if type(row[1]) is str:
            computed_topics = find_topics(row[1], relevant_topics_cat if relevant_topics_cat!=None else [])
            line.append(list(computed_topics))
        else:
            line.append([])
        #vedo se la descrizione contiene brand o youtubers
        if type(row[2]) is str:
            cited_brand, cited_youtubers = find_brands_and_youtubers(row[2])
            line.append(list(cited_brand))
            line.append(list(cited_youtubers))
        else:
            line.append([])
            line.append([])
        df_arr_computed.append(line)
    c =[item for item in df.columns]
    c.append('topics')
    c.append('brands')
    c.append('youtubers')
    df_computed = pd.DataFrame(df_arr_computed, columns = c)
    return df_computed

df_computed_all = compute_details(df_videos_details)

df_computed_all

df_computed_reduced = df_computed_all[df_computed_all.columns[[0,3,6,7,8]]]
df_computed_reduced

df_computed_all.to_csv('data/get_videos_details/videos_details_computed_complete.csv', index = False)
df_computed_reduced.to_csv('data/get_videos_details/videos_details_computed.csv', index = False)
