==> 4_get_edges.ipynb <==
import pandas as pd

import os

import numpy as np

import networkx as nx

import matplotlib.pyplot as plt

from IPython.display import clear_output

df_comments = pd.read_csv('data/get_comment_threads_output/comments_thread_sentiment_reduced_5.csv')
df_videos = pd.read_csv('data/get_videos_details/videos_details_computed.csv')
df_youtuber_videos = pd.read_csv('data/get_youtuber_videos_output/youtuber_videos.csv')
df_youtubers = pd.read_csv('data/get_youtubers_output/youtubers.csv')

df_videos

user_name_set= set()
nan_number=0
for comment in df_comments.values:
    if comment[2]!= None:
        user_name_set.add(comment[2])
    else:
        nan_number+=1

df_comments[df_comments.columns[2]].isnull().sum(axis = 0)#buono!

youtuber_video_join=df_youtubers.set_index('channel_id').join(df_youtuber_videos.set_index('channel_id'), on='channel_id',lsuffix='_yt', rsuffix='_ytvd')

youtuber_video_join = youtuber_video_join[youtuber_video_join.columns[[0,2]]]

youtuber_video_join = youtuber_video_join.set_index('video_id').join(df_videos.set_index('video_id'),on='video_id',lsuffix='_yt', rsuffix='_ytvd')

youtuber_video_join

df_comments

edges = []

percentili = {0.1:0,0.2:0,0.3:0,0.4:0,0.5:0,0.6:0,0.7:0,0.8:0,0.9:0,1.0:0}

percentili_edges = {0.1:[],0.2:[],0.3:[],0.4:[],0.5:[],0.6:[],0.7:[],0.8:[],0.9:[],1.0:[]}

saves = 0

for idx_row_1, row_1 in enumerate(zip(youtuber_video_join.index,youtuber_video_join.values)):
    print('Computing {}/{}'.format(idx_row_1,len(youtuber_video_join.index)))
    if idx_row_1 != 0 and idx_row_1%200==0:
        print('Saving files')
        np.save('data/get_edges/edges_{}.npy'.format(saves), edges)
        np.save('data/get_edges/percentile/percentili_{}.npy'.format(saves), percentili)
        for idx, percentile in enumerate(percentili_edges.keys()):
            np.save('data/get_edges/percentile/percentili_edges_{}_{}.npy'.format(saves,idx), percentili_edges[percentile])
        saves = saves + 1
        percentili = {0.1:0,0.2:0,0.3:0,0.4:0,0.5:0,0.6:0,0.7:0,0.8:0,0.9:0,1.0:0}
        percentili_edges = {0.1:[],0.2:[],0.3:[],0.4:[],0.5:[],0.6:[],0.7:[],0.8:[],0.9:[],1.0:[]}
        edges = []
    row_1_idx = row_1[0]
    row_1_val = row_1[1]
    matches_topics=set()
    matches_brands = set()
    matches_youtubers = set()
    if type(row_1_val[2]) is str:
        [matches_topics.add(item.strip()[1:-1]) for item in row_1_val[2][1:-1].split(',') if item.strip()[1:-1]!=""]
    if type(row_1_val[3]) is str:
        [matches_brands.add(item.strip()[1:-1]) for item in row_1_val[3][1:-1].split(',') if item.strip()[1:-1]!=""]
    if type(row_1_val[4]) is str:
        [matches_youtubers.add(item.strip()[1:-1]) for item in row_1_val[4][1:-1].split(',') if item.strip()[1:-1]!=""]
    matches_youtubers.add(row_1_val[0])
    rows_0 = (df_comments.loc[df_comments['video_id'] == row_1_idx]).values
    for idx_row_2, row_2 in enumerate(zip(youtuber_video_join.index[idx_row_1:],youtuber_video_join.values[idx_row_1:])):
        row_2_idx = row_2[0]
        row_2_val = row_2[1]
        details = {'youtubers':[],'topics':[],'brands':[], 'sentiment':0}
        topics_=set()
        brands_ = set()
        youtubers_ = set()
        if type(row_2_val[2]) is str:
            [topics_.add(item.strip()[1:-1]) for item in row_2_val[2][1:-1].split(',') if item.strip()[1:-1]!=""]
        if type(row_2_val[3]) is str:
            [brands_.add(item.strip()[1:-1]) for item in row_2_val[3][1:-1].split(',') if item.strip()[1:-1]!=""]
        if type(row_2_val[4]) is str:
            [youtubers_.add(item.strip()[1:-1]) for item in row_2_val[4][1:-1].split(',') if item.strip()[1:-1]!=""]
        youtubers_.add(row_2_val[0])

        common_topics = matches_topics.intersection(topics_)
        common_brands = matches_brands.intersection(brands_)
        common_youtubers = matches_youtubers.intersection(youtubers_)

        union_topics = matches_topics.union(topics_)
        union_brands = matches_brands.union(brands_)
        union_youtubers = matches_youtubers.union(youtubers_)
        
        w = (len(common_topics)+len(common_brands)+len(common_youtubers))/(len(union_topics)+len(union_brands)+len(union_youtubers))
        if w > 0:
            if len(common_topics)>0:
                details['topics'] = list(common_topics)
            if len(common_brands)>0:
                details['brands'] = list(common_brands)
            if len(common_youtubers)>0:
                details['youtubers'] = list(common_youtubers)
            rows_1 = (df_comments.loc[df_comments['video_id'] == row_2_idx]).values
            for id_c_0,comment_0 in enumerate(rows_0):
                for id_c_1,comment_1 in enumerate(rows_1):
                    if comment_0[2] != comment_1[2]:
                        if comment_0[3] == comment_1[3]:
                            if comment_0[3] == 1:
                                details['sentiment'] = 1
                            else:
                                details['sentiment'] = 0
                        else:
                            details['sentiment'] = -1
                        e = [(comment_0[2],comment_1[2],w),details]
                        edges.append(e)
                        for percentile in percentili_edges.keys():
                            if w < percentile and w>=(percentile-0.1):
                                percentili[percentile] += 1
                                percentili_edges[percentile].append(e)
                            if w == 1.0:
                                percentili[1.0] += 1
                                percentili_edges[1.0].append(e)
    clear_output(wait=True)
    
np.save('data/get_edges/edges_{}.npy'.format(saves), edges)
np.save('data/get_edges/percentile/percentili_{}.npy'.format(saves), percentili)
for idx, percentile in enumerate(percentili_edges.keys()):
    np.save('data/get_edges/percentile/percentili_edges_{}_{}.npy'.format(saves,idx), percentili_edges[percentile])
