==> 4_get_network.ipynb <==
import pandas as pd

import os

import numpy as np

import networkx as nx

import matplotlib.pyplot as plt

from IPython.display import clear_output

file_index = 0

df_comments = pd.read_csv('data/get_comment_threads_output/comments_thread_sentiment_reduced_{}.csv'.format(file_index))

df_comments

df_videos = pd.read_csv('data/get_videos_details/videos_details_computed.csv')

df_youtuber_videos = pd.read_csv('data/get_youtuber_videos_output/youtuber_videos.csv')

df_youtubers = pd.read_csv('data/get_youtubers_output/youtubers.csv')

first_graph = nx.DiGraph()

youtubers_name_set=set()
for youtuber in df_youtubers.values:
    youtubers_name_set.add(youtuber[0])

user_name_set= set()
nan_number=0
for comment in df_comments.values:
    if comment[2]!= None:
        user_name_set.add(comment[2])
    else:
        nan_number+=1

df_comments[df_comments.columns[2]].isnull().sum(axis = 0)#buono!

tot_nodes= list(youtubers_name_set.union(user_name_set))

#tolgo il nan
tot_nodes = tot_nodes[1:]

first_graph.add_nodes_from(tot_nodes)

youtuber_video_join=df_youtubers.set_index('channel_id').join(df_youtuber_videos.set_index('channel_id'), on='channel_id',lsuffix='_yt', rsuffix='_ytvd')

youtuber_video_join = youtuber_video_join[youtuber_video_join.columns[[0,2]]]

youtuber_video_join = youtuber_video_join.set_index('video_id').join(df_videos.set_index('video_id'),on='video_id',lsuffix='_yt', rsuffix='_ytvd')

youtuber_video_join

youtuber_comments = youtuber_video_join.join(df_comments.set_index('video_id'),on='video_id',lsuffix='_ytvd', rsuffix='_comm')

youtuber_comments  = (youtuber_comments[youtuber_comments.columns[[0,2,3,4,6,7]]]).dropna()
youtuber_comments

edges = []

#{'video_category':[row[1]],'topics':[row[2]],'brands':[row[3]],'youtubers':row[4],'sentiments':[row[6]],'mean_sentiment':np.mean([row[6]])}
#i duplicati li guardo quando metto gli edge del grafo (se esiste già l'edge, aggiungo altri dettagli)
#il mean sentiment lo valuterò dopo
i = 1
for idx_for,row_1 in enumerate(youtuber_comments.values):
    print('Comment: {}/{}'.format(idx_for,len(youtuber_comments.values)))
    clear_output(wait=True)
    matches_topics = row_1[1]
    matches_brands = row_1[2]
    matches_youtubers = row_1[3]
    for row_2 in youtuber_comments.values[i:]:
        link = False
        details = {'youtubers':[],'topics':[],'brands':[]}
        common_topics = []
        common_brands = []
        common_youtubers = []
        if any(x in row_2[1] for x in matches_topics):
            link = True
            details['topics'].append(row_2[1])
        if any(x in row_2[2] for x in matches_brands):
            link = True
            details['brands'].append(row_2[2])
        if any(x in row_2[3] for x in matches_youtubers):
            link = True
            details['youtubers'].append(row_2[3])
        if row_1[0] == row_2[0]:
            link = True
            details['youtubers'].append(row_2[0])
        if link:
            edges.append([(row_1[4],row_2[4]),details])
    i += 1

edges
