==> 6_selecting_sentiment.ipynb <==
import os

import numpy as np

import networkx as nx

import matplotlib.pyplot as plt

from IPython.display import clear_output

complete_graph = nx.read_gpickle('data/graphs/complete_graph.gpickle')

positive_graph = nx.Graph()
negative_graph = nx.Graph()
positive_gephi_graph = nx.Graph()
negative_gephi_graph = nx.Graph()

positive_graph.add_nodes_from(complete_graph.nodes(data=True))
negative_graph.add_nodes_from(complete_graph.nodes(data=True))
positive_gephi_graph.add_nodes_from(complete_graph.nodes(data=True))
negative_gephi_graph.add_nodes_from(complete_graph.nodes(data=True))

for key_edge, value_edge in zip(complete_graph.edges.keys(),complete_graph.edges.values()):
    #entrambi negativi
    if value_edge['sentiment'] == 0:
        negative_graph.add_edge(key_edge[0],key_edge[1],weight=value_edge['weight'],youtubers=value_edge['youtubers'], topics = value_edge['topics'], brands=value_edge['brands'])
        negative_gephi_graph.add_edge(key_edge[0],key_edge[1],weight=value_edge['weight'])
    #entrambi positivi
    if value_edge['sentiment'] == 1:
        positive_graph.add_edge(key_edge[0],key_edge[1],weight=value_edge['weight'],youtubers=value_edge['youtubers'], topics = value_edge['topics'], brands=value_edge['brands'])
        positive_gephi_graph.add_edge(key_edge[0],key_edge[1],weight=value_edge['weight'])

nx.write_gpickle(positive_graph,'data/graphs/positive_graph.gpickle')
nx.write_gpickle(negative_graph,'data/graphs/negative_graph.gpickle')
nx.write_gpickle(positive_gephi_graph,'data/graphs/positive_gephi_graph.gpickle')
nx.write_gpickle(negative_gephi_graph,'data/graphs/negative_gephi_graph.gpickle')

nx.write_gexf(positive_gephi_graph,'data/graphs/positive_gephi_graph.gexf')
nx.write_gexf(negative_gephi_graph,'data/graphs/negative_gephi_graph.gexf')
