==> 8_negative_get_famous_topics.ipynb <==
import pandas as pd

import os

import numpy as np

import networkx as nx

import matplotlib.pyplot as plt

from IPython.display import clear_output
from collections import Counter
import xlsxwriter

#importo il grafo completo
first_graph = nx.read_gpickle('data/graphs/negative/negative_graph.gpickle')

isolates = list(nx.isolates(first_graph))

#rimuovo i singoletti
first_graph.remove_nodes_from(list(nx.isolates(first_graph)))

topics = list(dict((np.load('data/topics/topic_ids.npy',allow_pickle=True)).tolist()).values())
[topics.append(item[1]) for item in np.load('data/topics/keywords.npy', allow_pickle=True).tolist()]
[topics.append(item) for item in list(dict((np.load('data/topics/categories_ids.npy',allow_pickle=True)).tolist()).values())]
topic_set = set(topics)
topic_set

brands = [item[-1] for item in np.load('data/topics/makeup_brands.npy', allow_pickle=True).tolist() if item[-1]!=""]
brand_set = set(brands)
brand_set

df_youtubers = pd.read_csv('data/get_youtubers_output/youtubers.csv')
youtubers_set = set(df_youtubers['display_name'].values)
youtubers_set

#preparo il counter
dict_attrs = dict()
for idx, attrs in enumerate(first_graph.nodes.values()):
    for attr in attrs.keys():
        dict_attrs.update({attr:0})

for idx, attrs in enumerate(first_graph.nodes.values()):
    print('{}/{}'.format(idx,len(first_graph.nodes)))
    clear_output(wait=True)
    for key_attr, key_value in zip (attrs.keys(),attrs.values()):
        dict_attrs[key_attr] += key_value

rank_topics=[]
rank_brands=[]
rank_youtubers=[]
for key_item,key_value in zip (dict_attrs.keys(),dict_attrs.values()):
    if key_item in topic_set:
        rank_topics.append((key_item,key_value))
    if key_item in brand_set:
        rank_brands.append((key_item,key_value))
    if key_item in youtubers_set:
        rank_youtubers.append((key_item,key_value))

rank_topics = sorted(rank_topics, key=lambda x:x[1],reverse = True)
rank_youtubers = sorted(rank_youtubers, key=lambda x:x[1],reverse = True)
rank_brands = sorted(rank_brands, key=lambda x:x[1],reverse = True)

print("Top 3 topics: {}".format(rank_topics[:3]))
print("Top 3 youtubers: {}".format(rank_youtubers[:3]))
print("Top 3 brands: {}".format(rank_brands[:3]))

assortativity_coefficients = np.load('analysis/negative/assortativity_coefficients.npy',allow_pickle=True)
assortativity_coefficients=dict(assortativity_coefficients.tolist())

print("TOPICS RANKING")
rows=[]
for idx, t in enumerate(rank_topics):
    rows.append((idx+1,t[0],assortativity_coefficients[t[0]]))
    
df =pd.DataFrame(rows, columns=["Position","Topic","Assortativity coefficient"])
df

print("BRANDS RANKING")
rows=[]
for idx, t in enumerate(rank_brands):
    rows.append((idx+1,t[0],assortativity_coefficients[t[0]]))
    
df =pd.DataFrame(rows, columns=["Position","Brand","Assortativity coefficient"])
df

print("YOUTUBERS RANKING")
rows=[]
for idx, t in enumerate(rank_youtubers):
    rows.append((idx+1,t[0],assortativity_coefficients[t[0]]))
    
df =pd.DataFrame(rows, columns=["Position","Youtuber","Assortativity coefficient"])
df

## Famous topic communities

partition_louvain = np.load("analysis/negative/partition_louvain.npy", allow_pickle=True).tolist()

community_preferences_counter=dict()
community_preferences=dict()

for c in np.unique(list(partition_louvain.values())):
    dict_attrs = dict()
    for idx, attrs in enumerate(first_graph.nodes.values()):
        for attr in attrs.keys():
            dict_attrs.update({attr:0})
        break
    insert = False
    for key, val in zip(list(partition_louvain.keys()),list(partition_louvain.values())):
        if val == c and key not in isolates:
            insert = True
            for key_attr, key_value in zip (first_graph.nodes[key].keys(),first_graph.nodes[key].values()):
                dict_attrs[key_attr] += key_value
    if insert:
        community_preferences_counter.update({c:dict_attrs})

for c in community_preferences_counter:
    rank_topics=[]
    rank_brands=[]
    rank_youtubers=[]
    for key_item,key_value in zip (community_preferences_counter[c].keys(),community_preferences_counter[c].values()):
        if key_item in topic_set:
            rank_topics.append((key_item,key_value))
        if key_item in brand_set:
            rank_brands.append((key_item,key_value))
        if key_item in youtubers_set:
            rank_youtubers.append((key_item,key_value))
    rank_topics = sorted(rank_topics, key=lambda x:x[1],reverse = True)
    rank_youtubers = sorted(rank_youtubers, key=lambda x:x[1],reverse = True)
    rank_brands = sorted(rank_brands, key=lambda x:x[1],reverse = True)
    community_preferences.update({c:{"youtubers":rank_youtubers[:3],"topics":rank_topics[:3],"brands":rank_brands[:3]}})

counter = Counter(partition_louvain.values())
for c in community_preferences:
    print("__________________________________________________")
    print("COMMUNITY {} with {} users".format(c,counter[c]))
    print("Favourite youtubers: {}".format([item[0] for item in community_preferences[c]['youtubers'] if item[1]!=0]))
    print("Favourite topics: {}".format([item[0] for item in community_preferences[c]['topics'] if item[1]!=0]))
    print("Favourite brands: {}".format([item[0] for item in community_preferences[c]['brands'] if item[1]!=0]))

## Possible overlapping communities

overlapping_matrix= []
for c_1 in community_preferences:
    yt_1 = [item[0] for item in community_preferences[c_1]['youtubers'] if item[1]!=0]
    tp_1 = [item[0] for item in community_preferences[c_1]['topics'] if item[1]!=0]
    br_1 = [item[0] for item in community_preferences[c_1]['brands'] if item[1]!=0]
    overlapping_matrix.append([])
    for c_2 in community_preferences:
        yt_2 = [item[0] for item in community_preferences[c_2]['youtubers'] if item[1]!=0]
        tp_2 = [item[0] for item in community_preferences[c_2]['topics'] if item[1]!=0]
        br_2 = [item[0] for item in community_preferences[c_2]['brands'] if item[1]!=0]
        over = (len(set(yt_1).intersection(set(yt_2)))+len(set(tp_1).intersection(set(tp_2)))+len(set(br_1).intersection(set(br_2))))/(len(set(yt_1).union(set(yt_2)))+len(set(tp_1).union(set(tp_2)))+len(set(br_1).union(set(br_2))))
        overlapping_matrix[-1].append(over)

keys = community_preferences.keys()
workbook = xlsxwriter.Workbook('analysis/negative/overlapping_matrix.xlsx') 
worksheet = workbook.add_worksheet('Overlapping matrix') 
currency_format = workbook.add_format({'num_format': '0.00'})
cell_format_yellow = workbook.add_format()
cell_format_yellow.set_bg_color('yellow')
cell_format_yellow.set_num_format('0.00')
cell_format_red = workbook.add_format()
cell_format_red.set_bg_color('red')
cell_format_red.set_num_format('0.00')
for idx,val in enumerate(keys):
    worksheet.write(0, idx+1,val)
    worksheet.write(idx+1,0,val)
for row_idx, row in enumerate(overlapping_matrix):
    for col_idx, col in enumerate(overlapping_matrix):
        if overlapping_matrix[row_idx][col_idx] >=0.5 and overlapping_matrix[row_idx][col_idx] < 0.8:
            worksheet.write(row_idx+1, col_idx+1,overlapping_matrix[row_idx][col_idx],cell_format_yellow)
        elif overlapping_matrix[row_idx][col_idx] >=0.8:
            worksheet.write(row_idx+1, col_idx+1,overlapping_matrix[row_idx][col_idx],cell_format_red)
        else:
            worksheet.write(row_idx+1, col_idx+1,overlapping_matrix[row_idx][col_idx],currency_format)
workbook.close()
